+++ tsunami-udp/rtserver/transcript.c
@@ -153,7 +153,7 @@ void xscript_open(ttp_session_t *session)
     fprintf(xfer->transcript, "faster_num = %u\n",  param->faster_num);
     fprintf(xfer->transcript, "faster_den = %u\n",  param->faster_den);
     fprintf(xfer->transcript, "ipd_time = %u\n",    param->ipd_time);
-    fprintf(xfer->transcript, "ipd_current = %u\n", xfer->ipd_current);
+    fprintf(xfer->transcript, "ipd_current = %u\n", (u_int32_t)xfer->ipd_current);
     fprintf(xfer->transcript, "protocol_version = 0x%x\n", PROTOCOL_REVISION);
     fprintf(xfer->transcript, "software_version = %s\n",   TSUNAMI_CVS_BUILDNR);
     fprintf(xfer->transcript, "ipv6 = %u\n",        param->ipv6_yn);
@@ -164,6 +164,9 @@ void xscript_open(ttp_session_t *session)
 
 /*========================================================================
  * $Log$
+ * Revision 1.5  2007/12/07 18:10:28  jwagnerhki
+ * cleaned away 64-bit compile warnings, used tsunami-client.h
+ *
  * Revision 1.4  2007/10/29 15:30:25  jwagnerhki
  * timeout feature for rttsunamid too, added version info to transcripts, added --hbimeout srv cmd line param
  *
